requirements.txt
asgi-redis==1.4.3
asgiref==1.1.2
astroid==1.6.4
async-timeout==2.0.1
atomicwrites==1.1.5
attrs==21.2.0
autobahn==18.5.1
Automat==20.2.0
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.11.1
bleach==2.1.3
boto==2.49.0
boto3==1.14.15
botocore==1.17.15
certifi==2019.9.11
channels==1.1.8
chardet==3.0.4
charset-normalizer==2.1.1
colorama==0.3.9
configparser==3.5.0
constantly==15.1.0
coverage==4.5.1
daphne==1.4.2
decorator==4.3.0
defusedxml==0.6.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.6
django-allauth==0.41.0
django-appconf==1.0.3
django-autoslug==1.9.3
django-bmemcached==0.3.0
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap-datetimepicker==1.5
django-bootstrap-modal-forms==1.5.0
django-bootstrap3==11.1.0
django-bootstrap3-datetimepicker-2==2.4.2
django-bootstrap4==22.2
django-cities==0.5.0.4
django-ckedit==1.0
django-ckeditor==5.2.2
django-constrainedfilefield==4.0.0
django-coverage-plugin==1.5.0
django-crispy-forms==1.7.2
django-datetime-widget==0.9.3
django-debug-toolbar==1.9.1
django-el-pagination==3.2.4
django-friendship==1.5.0
django-fs-trumbowyg==0.1.4
django-localflavor==1.0
django-markdown-deux==1.0.5
django-markdownx==3.0.1
django-messages==0.5.3
django-meta==1.7.0
django-model-utils==3.0.0
django-modeltranslation==0.15.2
django-pagedown==2.1.2
django-quilljs==0.0.2
django-registration-redux==2.4
django-robots==4.0
django-select2==7.1.1
django-seo==0.3.5
django-storages==1.9.1
django-summernote==0.8.11.6
django-taggit==0.22.2
django-test-plus==1.0.22
django-utils-six==2.0
django-widget-tweaks==1.4.5
docutils==0.15.2
easy-thumbnails==2.8.3
gunicorn==20.0.4
html5lib==1.1
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.12.0
incremental==21.3.0
jmespath==0.10.0
lazy-object-proxy==1.3.1
Markdown==3.4.1
markdown2==2.4.3
more-itertools==8.14.0
msgpack-python==0.5.6
oauthlib==3.1.0
pathlib==1.0.1
pathlib2==2.3.2
Pillow==9.2.0
pluggy==0.6.0
psycopg2==2.8.5
py==1.11.0
pytest==3.6.0
pytest-cov==2.5.1
pytest-django==2.9.1
pytest-sugar==0.9.1
python-binary-memcached==0.30.0
python-dateutil==2.7.3
python-decouple==3.1
python-dotenv==0.14.0
python-gettext==4.0
python3-openid==3.1.0
pytz==2018.4
redis==2.10.6
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.3.7
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
static3==0.7.0
swapper==1.3.0
termcolor==1.1.0
tqdm==4.64.1
Twisted==22.8.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.3.0
uhashring==2.1
Unipath==1.1
urllib3==1.25.3
virtualenv==15.2.0
webencodings==0.5.1
whitenoise==3.3.1
wrapt==1.14.1
zipp==3.8.1
zope.interface==5.4.0
